requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorclass==2.2.2
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.57.0
django-allauth-ui==0.1.4
django-classy-tags==4.1.0
django-environ==0.11.2
django-sekizai==4.1.0
django-widget-tweaks==1.5.0
docopt==0.6.2
idna==3.4
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pip-upgrader==1.4.15
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
terminaltables==3.1.10
typing_extensions==4.8.0
urllib3==2.0.7
